+++ NEWS.md
@@ -1,5 +1,8 @@
 # resourcecode (development version)
 
+- New function 'weather_windows()' to compute weather windows of a given duration.
+- Update the ending date of the Hindcast to 2024-12-31.
+
 # resourcecode 0.4.0
 
 - New example data: `rscd_data_example` with an extract of some variables at a given node.